Add files via upload
This dataset is made up of 12 observed variables and 560 observations. It was collected via Digi company, through the survey. It was meant for confirmatory composite analysis (CCA) as 3 emergent variables had emerged from this dataset. The emerging models are "access on the internet", "privacy issue on the internet", "social influence on the internet".